Rewrite editor internals to use context
This is a large overhaul of the editor internal state to
make things easier to reason about. The primary editor code
was hacked together quickly, with asynchronous effects updating
state that should have been serial. It also removes a lot of
prop drilling which was a burden to follow.

This now breaks app state into its primary logical parts:

- **ThemeProvider**: The app's internal Theme UI config
- **BlocksProvider**: The renderable block components passed to
  the editor
- **EditorProvider**: The editor state including active tab,
  mode, etc.
 - **ElementProvider**: The current selected element and its metadata
- **ThemeEditorProvider**: The theme config for the canvas
- **ScopeProvider**: The components and properties that the canvas
  has access to
- **CodeProvider**: The current JSX code, its transformed version,
  and other data derived from the JSX that the editor UI needs. This
  also has the numerous updaters that transform the code with babel
  plugins.